Add pundit

+++ app/controllers/midgets_controller.rb
@@ -3,7 +3,7 @@ class MidgetsController < ApplicationController
   skip_before_action :authenticate_user!, only: %i[index show]
 
   def index
-    @midgets = Midget.all
+    @midgets = policy_scope(Midget).order(created_at: :desc)
 
     @markers = @midgets.geocoded.map do |midget|
       {
@@ -17,11 +17,13 @@ class MidgetsController < ApplicationController
 
   def new
     @midget = Midget.new
+    authorize @midget
   end
 
   def create
     @midget = Midget.new(midget_params)
     @midget.user_id = current_user.id
+    authorize @midget
     if @midget.save
       redirect_to midget_path(@midget)
     else
@@ -30,6 +32,7 @@ class MidgetsController < ApplicationController
   end
 
   def show
+    authorize @midget
   end
 
   def edit
@@ -44,6 +47,8 @@ class MidgetsController < ApplicationController
   end
 
   def destroy
+    authorize @midget
+
     if @midget.destroy
       redirect_to midgets_path
     else
